requirements.txt
Django==1.4.1
django-sitetree==0.9
chunks==0.1
wsgiref==0.1.2
pil==1.1.7
django-markdown==0.1.0
django-haystack==1.2.7
pysolr
django-tastypie==0.9.11